Fix MSVC warnings

+++ reservoir_sampler.h
@@ -26,8 +26,8 @@
 #include <cassert>
 #include <cstring>
 #include <random>
-#include <vector>
 #include <type_traits>
+#include <vector>
 
 template<typename T, typename URBG = std::mt19937, typename RandType = float>
 class ReservoirSampler
@@ -170,7 +170,7 @@ private:
 			if (mFilledElementsCount == mSamplesCount)
 			{
 				mWeightJumpOver = std::exp(std::log(mUniformDist(mRand)) / mSamplesCount);
-				mIndexesToJumpOver = std::floor(std::log(mUniformDist(mRand))/std::log(static_cast<RandType>(1.0) - mWeightJumpOver));
+				mIndexesToJumpOver = static_cast<size_t>(std::floor(std::log(mUniformDist(mRand))/std::log(static_cast<RandType>(1.0) - mWeightJumpOver)));
 			}
 		}
 		else
@@ -180,7 +180,7 @@ private:
 				replaceElement<isT>(std::forward<Args>(arguments)...);
 
 				mWeightJumpOver *= std::exp(std::log(mUniformDist(mRand)) / mSamplesCount);
-				mIndexesToJumpOver += std::floor(std::log(mUniformDist(mRand))/std::log(static_cast<RandType>(1.0) - mWeightJumpOver));
+				mIndexesToJumpOver += static_cast<size_t>(std::floor(std::log(mUniformDist(mRand))/std::log(static_cast<RandType>(1.0) - mWeightJumpOver)));
 			}
 			else
 			{
